Rename hex_color module to avoid warning
