+++ crates/ecolor/src/hex_color_macro.rs
